requirements.txt
# ToxiGen Hate Speech Detection Project
# Core dependencies
datasets>=2.0.0
torch>=1.9.0
transformers>=4.20.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Data processing
pyarrow>=5.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Prompt validation and Azure AI
azure-ai-inference>=1.0.0b9

# Optional: for better performance
accelerate>=0.20.0